# DBZ-7421 Improve incremental snapshot performance with increasing number of collections to snapshot