a possibile solution for https://issues.redhat.com/browse/DBZ-7421

Before
```
Benchmark                             (numberOfTableToSnapshot)   Mode  Cnt       Score   Error  Units
IncrementalSnapshotContextPerf.store                          1  thrpt    2  962999.691          ops/s
IncrementalSnapshotContextPerf.store                          5  thrpt    2  118903.880          ops/s
IncrementalSnapshotContextPerf.store                         10  thrpt    2   31712.744          ops/s
IncrementalSnapshotContextPerf.store                         50  thrpt    2     786.632          ops/s
IncrementalSnapshotContextPerf.store                        100  thrpt    2     195.274          ops/s
```

After some code optimization
```
Benchmark                             (numberOfTableToSnapshot)   Mode  Cnt       Score   Error  Units
IncrementalSnapshotContextPerf.store                          1  thrpt    2  926594.940          ops/s
IncrementalSnapshotContextPerf.store                          5  thrpt    2  130743.643          ops/s
IncrementalSnapshotContextPerf.store                         10  thrpt    2   30547.231          ops/s
IncrementalSnapshotContextPerf.store                         50  thrpt    2     873.143          ops/s
IncrementalSnapshotContextPerf.store                        100  thrpt    2     222.739          ops/s
```

With the cache
```
Benchmark                             (numberOfTableToSnapshot)   Mode  Cnt        Score   Error  Units
IncrementalSnapshotContextPerf.store                          1  thrpt    2  1794620.143          ops/s
IncrementalSnapshotContextPerf.store                          5  thrpt    2  1984367.094          ops/s
IncrementalSnapshotContextPerf.store                         10  thrpt    2  1792400.614          ops/s
IncrementalSnapshotContextPerf.store                         50  thrpt    2  1985691.291          ops/s
IncrementalSnapshotContextPerf.store                        100  thrpt    2  1911970.545          ops/s
```

Even if improving the code that serializes the data collections to JSON, the performance was not good. Adding a cache seems the best approach here. 
With this implementation the serialize code will not affect the `store` method that is called during the event `IncrementalSnapshotChangeRecordReceiver` but will affect the methods that modify the `dataCollectionsToSnapshot` variable, but the impact in this case is limited to the number of data collections to snapshot and not at every events. 

@Naros, @jpechane @vjuranek @jcechace  WDYT?