@jpechane 2024-01-29
@mfvitale I think the solution seems to be right! I'd just prefer now as both the attributes are linked together not to manipulate them directly in the methods but instead provide specific methods that will change both of them when needed. WDYT?