@jpechane 2024-01-30
@mfvitale I think it is good, thanks!
Theoretically it should be delegation (HAS-A) not inheritance (IS-A) relationship but I think it is good pragmatic solution.